Added a link to view the Attendance Policy PDF Document.
NOTE TO UI TEAM - the link text was invisible - displayed as white and blended with the background image. I changed the Site.css and FAQ Index.cshtml, adding a CSS class called   darklink   changing the link text to Black, so it would be visible.  Make changes to both these files as you see fit!   (John Warner)